+++ examples/tsmodels.py
@@ -4,7 +4,7 @@ from stdnet.apps.columnts import ColumnTSField
 
         
 class TimeSeries(odm.StdModel):
-    ticker = odm.SymbolField(unique = True)
+    ticker = odm.SymbolField(unique=True)
     data  = odm.TimeSeriesField()
     
     def todate(self, v):
@@ -38,23 +38,22 @@ from which data is not available. It return a list of two-dimensional tuples
 containing start and end date for the interval. The list could countain 0,1 or 2
 tuples.'''
         return missing_intervals(startdate, enddate, self.data_start,
-                                 self.data_end, dateconverter = self.todate,
-                                 parseinterval = parseinterval)
+                                 self.data_end, dateconverter=self.todate,
+                                 parseinterval=parseinterval)
     
 
 class DateTimeSeries(TimeSeries):
-    data  = odm.TimeSeriesField(pickler = encoders.DateConverter())
+    data  = odm.TimeSeriesField(pickler=encoders.DateConverter())
     
     def todate(self, v):
         return todate(v)
     
     
 class BigTimeSeries(DateTimeSeries):
-    data  = odm.TimeSeriesField(
-                    pickler = encoders.DateConverter(),
-                    value_pickler = encoders.PythonPickle())
+    data  = odm.TimeSeriesField(pickler=encoders.DateConverter(),
+                                value_pickler=encoders.PythonPickle())
 
 
 class ColumnTimeSeries(odm.StdModel):
-    ticker = odm.SymbolField(unique = True)
+    ticker = odm.SymbolField(unique=True)
     data = ColumnTSField()